Feat(Login): Created the login ui still needs functionality
