+++ src/app/(pages)/Auth/Login/page.tsx
@@ -1,12 +1,38 @@
-import { LoginForm } from "@/components/login-form";
-import React from "react";
+import { AudioLines } from "lucide-react";
+import Image from "next/image";
 
-const LoginPage = () => {
+import { LoginForm } from "@/components/login-form";
+
+export default function LoginPage() {
   return (
-    <div>
-      <LoginForm />
+    <div className="grid min-h-svh lg:grid-cols-2">
+      <div className="flex flex-col gap-4 p-6 md:p-10">
+        <div className="flex justify-center gap-2 md:justify-start">
+          <a
+            href="#"
+            className="flex items-center gap-2 font-medium --font-poppins"
+          >
+            <div className="flex h-6 w-6 items-center justify-center rounded-md bg-primary text-primary-foreground">
+              <AudioLines className="size-4" />
+            </div>
+            3legant
+          </a>
+        </div>
+        <div className="flex flex-1 items-center justify-center">
+          <div className="w-full max-w-xs">
+            <LoginForm />
+          </div>
+        </div>
+      </div>
+      <div className="relative hidden bg-muted lg:block">
+        <Image
+          src="/images/headphone-hero2.jpg"
+          alt="Image"
+          className="absolute inset-0 h-full w-full object-cover dark:brightness-[0.6] dark:grayscale"
+          width={1000}
+          height={1000}
+        />
+      </div>
     </div>
   );
-};
-
-export default LoginPage;
+}
